@@ -1,13 +1,17 @@
-name: scroll_app_bar_exemples
+name: scroll_app_bar_example
+publish_to: none
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   scroll_app_bar:
     path: ..
+
+dev_dependencies: 
+  flutter_lints: ^1.0.4
 
 flutter:
   uses-material-design: true
