@@ -1,17 +1,15 @@
 name: scroll_bars_common
 description: A common package to scroll bars packages. This package is not intended for direct use.
-version: 2.0.0
-homepage: https://github.com/EdsonOnildoJR/scroll_bars_common
+version: 3.0.0
+homepage: https://github.com/edsonbonfim/scroll_bars/tree/master/scroll_bars_common
+issue_tracker: https://github.com/edsonbonfim/scroll_bars/issues
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
-flutter:
+  flutter_lints: ^1.0.4
