@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:scroll_app_bar/scroll_app_bar.dart';
 
-class GradientPage extends StatelessWidget {
-  GradientPage({Key? key}) : super(key: key);
+class BasicPage extends StatelessWidget {
+  BasicPage({Key? key}) : super(key: key);
 
   final controller = ScrollController();
 
@@ -11,16 +11,7 @@ class GradientPage extends StatelessWidget {
     return Scaffold(
       appBar: ScrollAppBar(
         controller: controller, // Note the controller here
-        title: const Text("Gradient Background"),
-        backgroundGradient: const LinearGradient(
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-          colors: [
-            Colors.red,
-            Colors.green,
-            Colors.blue,
-          ],
-        ),
+        title: const Text("Basic"),
       ),
       body: ListView(
         controller: controller, // Controller is also here
