@@ -1,28 +1,42 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:scroll_bars_common/scroll_bars_common.dart';
-
-import 'package:flutter/foundation.dart' show kIsWeb;
 
 extension ScrollAppBarControllerExt on ScrollController {
   static final _controllers = <int, _ScrollAppBarController>{};
 
   _ScrollAppBarController get appBar {
-    if (_controllers.containsKey(this.hashCode)) {
-      return _controllers[this.hashCode];
+    if (_controllers.containsKey(hashCode) &&
+        _controllers[hashCode] is _ScrollAppBarController) {
+      return _controllers[hashCode]!;
     }
 
-    return _controllers[this.hashCode] = _ScrollAppBarController(this);
+    return _controllers[hashCode] = _ScrollAppBarController(this);
+  }
+
+  bottom({
+    double? toolbarHeight,
+    double? preferredSizeHeight,
+  }) {
+    _controllers[hashCode] = _ScrollAppBarController(
+      this,
+      toolbarHeight: toolbarHeight,
+      preferredSizeHeight: preferredSizeHeight,
+    );
   }
 }
 
 class _ScrollAppBarController extends ScrollBarsController {
-  _ScrollAppBarController(ScrollController scrollController)
-      : assert(scrollController != null),
+  _ScrollAppBarController(
+    ScrollController scrollController, {
+    this.toolbarHeight,
+    this.preferredSizeHeight,
+  })  : height = toolbarHeight ?? kToolbarHeight + (preferredSizeHeight ?? 0.0),
         super(scrollController);
 
+  final double? toolbarHeight;
+
+  final double? preferredSizeHeight;
+
   @override
-  double height =
-      kToolbarHeight + (kIsWeb ? 0.0 : Platform.isAndroid ? 24.0 : 0.0);
+  double height;
 }
