@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'page/home_page.dart';
 
 class App extends StatelessWidget {
+  const App({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: "Scroll bottom navigation bar",
+    return const MaterialApp(
+      title: "Scroll Bottom Navigation Bar",
       home: HomePage(),
-      themeMode: ThemeMode.dark,
-      darkTheme: ThemeData.dark(),
     );
   }
 }
