@@ -3,13 +3,10 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:scroll_bottom_navigation_bar/scroll_bottom_navigation_bar.dart';
 
-class PinUnpinPage extends StatefulWidget {
-  @override
-  _PinUnpinPageState createState() => _PinUnpinPageState();
-}
+class PinUnpinPage extends StatelessWidget {
+  PinUnpinPage({Key? key}) : super(key: key);
 
-class _PinUnpinPageState extends State<PinUnpinPage> {
-  ScrollController controller;
+  final controller = ScrollController();
 
   static const _items = <BottomNavigationBarItem>[
     BottomNavigationBarItem(
@@ -23,42 +20,20 @@ class _PinUnpinPageState extends State<PinUnpinPage> {
   ];
 
   @override
-  void initState() {
-    super.initState();
-    controller = ScrollController();
-  }
-
-  @override
-  void dispose() {
-    controller.bottomNavigationBar.dispose();
-    controller.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text("Snap scroll")),
+      appBar: AppBar(title: const Text("Basic")),
       body: ValueListenableBuilder<int>(
         valueListenable: controller.bottomNavigationBar.tabNotifier,
-        builder: (context, value, child) => Snap(
-          key: PageStorageKey(value),
-          controller: controller.bottomNavigationBar,
-          child: ListView.builder(
-            controller: controller, // Controller is also here
-            itemBuilder: (context, index) {
-              return Padding(
-                padding: const EdgeInsets.all(20.0),
-                child: Center(
-                  child: Text('$index'),
-                ),
-              );
-            },
-          ),
+        child: SizedBox(height: MediaQuery.of(context).size.height * 2),
+        builder: (context, pageIndex, child) => ListView(
+          key: PageStorageKey(pageIndex),
+          controller: controller, // Note the controller here
+          children: [child!],
         ),
       ),
       bottomNavigationBar: ScrollBottomNavigationBar(
-        controller: controller, // Note the controller here
+        controller: controller, // Controller is also here
         items: _items,
       ),
       floatingActionButton: FloatingActionButton(
@@ -71,9 +46,9 @@ class _PinUnpinPageState extends State<PinUnpinPage> {
         },
         child: ValueListenableBuilder<bool>(
           valueListenable: controller.bottomNavigationBar.pinNotifier,
-          child: Icon(Icons.push_pin),
+          child: const Icon(Icons.push_pin),
           builder: (context, value, child) {
-            if (!value) return child;
+            if (!value) return child!;
             return Transform.rotate(angle: pi / 2, child: child);
           },
         ),
