@@ -5,19 +5,17 @@ extension ScrollBottomNavigationBarControllerExt on ScrollController {
   static final _controllers = <int, _ScrollBottomNavigationBarController>{};
 
   _ScrollBottomNavigationBarController get bottomNavigationBar {
-    if (_controllers.containsKey(this.hashCode)) {
-      return _controllers[this.hashCode];
+    if (_controllers.containsKey(hashCode)) {
+      return _controllers[hashCode]!;
     }
 
-    return _controllers[this.hashCode] =
-        _ScrollBottomNavigationBarController(this);
+    return _controllers[hashCode] = _ScrollBottomNavigationBarController(this);
   }
 }
 
 class _ScrollBottomNavigationBarController extends ScrollBarsController {
   _ScrollBottomNavigationBarController(ScrollController scrollController)
-      : assert(scrollController != null),
-        super(scrollController);
+      : super(scrollController);
 
   @override
   double height = kBottomNavigationBarHeight;
